requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
distro==1.9.0
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tabula-py==2.9.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
